+++ features/config/new-branch-push-flag/display.feature
@@ -1,14 +1,14 @@
 Feature: display the new-branch-push-flag setting
 
   Scenario: default local setting
-    When I run "git-town new-branch-push-flag"
+    When I run "git-town config new-branch-push-flag"
     Then it prints:
       """
       false
       """
 
   Scenario: default global setting
-    When I run "git-town new-branch-push-flag --global"
+    When I run "git-town config new-branch-push-flag --global"
     Then it prints:
       """
       false
@@ -16,7 +16,7 @@ Feature: display the new-branch-push-flag setting
 
   Scenario Outline: local setting
     Given setting "new-branch-push-flag" is "<VALUE>"
-    When I run "git-town new-branch-push-flag"
+    When I run "git-town config new-branch-push-flag"
     Then it prints:
       """
       <VALUE>
@@ -28,7 +28,7 @@ Feature: display the new-branch-push-flag setting
 
   Scenario Outline: global setting
     Given setting "new-branch-push-flag" is globally "<VALUE>"
-    When I run "git-town new-branch-push-flag --global"
+    When I run "git-town config new-branch-push-flag --global"
     Then it prints:
       """
       <VALUE>
@@ -41,7 +41,7 @@ Feature: display the new-branch-push-flag setting
 
   Scenario: global set, local not set
     Given setting "new-branch-push-flag" is globally "true"
-    When I run "git-town new-branch-push-flag"
+    When I run "git-town config new-branch-push-flag"
     Then it prints:
       """
       true
@@ -50,7 +50,7 @@ Feature: display the new-branch-push-flag setting
   Scenario: global and local set
     Given setting "new-branch-push-flag" is globally "true"
     And setting "new-branch-push-flag" is "false"
-    When I run "git-town new-branch-push-flag"
+    When I run "git-town config new-branch-push-flag"
     Then it prints:
       """
       false
@@ -58,7 +58,7 @@ Feature: display the new-branch-push-flag setting
 
   Scenario: invalid value
     Given setting "new-branch-push-flag" is "zonk"
-    When I run "git-town new-branch-push-flag"
+    When I run "git-town config new-branch-push-flag"
     Then it prints:
       """
       Invalid value for git-town.new-branch-push-flag: "zonk". Please provide either true or false. Considering false for now.
