+++ features/config/new-branch-push-flag/update.feature
@@ -1,7 +1,7 @@
 Feature: set the new-branch-push-flag
 
   Scenario Outline: local setting
-    When I run "git-town new-branch-push-flag <GIVE>"
+    When I run "git-town config new-branch-push-flag <GIVE>"
     Then setting "new-branch-push-flag" is now "<WANT>"
 
     Examples:
@@ -14,14 +14,14 @@ Feature: set the new-branch-push-flag
       | 0     | false |
 
   Scenario: invalid value
-    When I run "git-town new-branch-push-flag zonk"
+    When I run "git-town config new-branch-push-flag zonk"
     Then it prints the error:
       """
       invalid argument: "zonk". Please provide either "true" or "false"
       """
 
   Scenario Outline: global setting
-    When I run "git-town new-branch-push-flag --global <GIVE>"
+    When I run "git-town config new-branch-push-flag --global <GIVE>"
     Then setting "new-branch-push-flag" is now "<WANT>"
 
     Examples:
