+++ features/help/help.feature
@@ -9,29 +9,29 @@ Feature: help for commands
       """
 
     Examples:
-      | COMMAND              |
-      | alias                |
-      | append               |
-      | completions          |
-      | config               |
-      | config main-branch   |
-      | diff-parent          |
-      | hack                 |
-      | help                 |
-      | kill                 |
-      | new-branch-push-flag |
-      | new-pull-request     |
-      | offline              |
-      | perennial-branches   |
-      | prepend              |
-      | prune-branches       |
-      | pull-branch-strategy |
-      | rename-branch        |
-      | repo                 |
-      | set-parent-branch    |
-      | ship                 |
-      | sync                 |
-      | version              |
+      | COMMAND                     |
+      | alias                       |
+      | append                      |
+      | completions                 |
+      | config                      |
+      | config main-branch          |
+      | config new-branch-push-flag |
+      | diff-parent                 |
+      | hack                        |
+      | help                        |
+      | kill                        |
+      | new-pull-request            |
+      | offline                     |
+      | perennial-branches          |
+      | prepend                     |
+      | prune-branches              |
+      | pull-branch-strategy        |
+      | rename-branch               |
+      | repo                        |
+      | set-parent-branch           |
+      | ship                        |
+      | sync                        |
+      | version                     |
 
   Scenario Outline: outside a Git repository
     Given I am outside a Git repo
