+++ src/cmd/config.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"strconv"
 
 	"github.com/git-town/git-town/v7/src/cli"
 	"github.com/git-town/git-town/v7/src/git"
@@ -54,6 +55,45 @@ func printMainBranch() {
 	cli.Println(cli.PrintableMainBranch(prodRepo.Config.MainBranch()))
 }
 
+var newBranchPushFlagCommand = &cobra.Command{
+	Use:   "new-branch-push-flag [(true | false)]",
+	Short: "Displays or sets your new branch push flag",
+	Long: `Displays or sets your new branch push flag
+
+If "new-branch-push-flag" is true, Git Town pushes branches created with
+hack / append / prepend on creation. Defaults to false.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		if len(args) == 0 {
+			printNewBranchPushFlag(prodRepo)
+		} else {
+			value, err := strconv.ParseBool(args[0])
+			if err != nil {
+				cli.Exit(fmt.Errorf(`invalid argument: %q. Please provide either "true" or "false"`, args[0]))
+			}
+			err = setNewBranchPushFlag(value, prodRepo)
+			if err != nil {
+				cli.Exit(err)
+			}
+		}
+	},
+	Args: cobra.MaximumNArgs(1),
+	PreRunE: func(cmd *cobra.Command, args []string) error {
+		return ValidateIsRepository(prodRepo)
+	},
+}
+
+func printNewBranchPushFlag(repo *git.ProdRepo) {
+	if globalFlag {
+		cli.Println(strconv.FormatBool(repo.Config.ShouldNewBranchPushGlobal()))
+	} else {
+		cli.Println(cli.PrintableNewBranchPushFlag(prodRepo.Config.ShouldNewBranchPush()))
+	}
+}
+
+func setNewBranchPushFlag(value bool, repo *git.ProdRepo) error {
+	return repo.Config.SetNewBranchPush(value, globalFlag)
+}
+
 func setMainBranch(branchName string, repo *git.ProdRepo) error {
 	hasBranch, err := repo.Silent.HasLocalBranch(branchName)
 	if err != nil {
@@ -101,6 +141,8 @@ var setupConfigCommand = &cobra.Command{
 
 func init() {
 	configCommand.AddCommand(mainBranchConfigCommand)
+	newBranchPushFlagCommand.Flags().BoolVar(&globalFlag, "global", false, "Displays or sets your global new branch push flag")
+	configCommand.AddCommand(newBranchPushFlagCommand)
 	configCommand.AddCommand(resetConfigCommand)
 	configCommand.AddCommand(setupConfigCommand)
 	RootCmd.AddCommand(configCommand)
